% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.beta.R
\name{bootstrap.y}
\alias{bootstrap.y}
\title{SAE bootstrap ys}
\usage{
bootstrap.y(model1, model_fit1, censusdata1, n_boot1, n_obs)
}
\arguments{
\item{x}{Beschreibung von \code{x}}

\item{y}{Beschreibung von \code{y}}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
