\name{optimizeRE}
\alias{optimizeRE}
\alias{optimizeRE.MSRFH}
\alias{optimizeRE.MSSTRFH}
\alias{optimizeRE.default}
\title{optimizeRE - generic funciton}
\usage{
  optimizeRE(modelSpecs)

  optimizeRE.MSSTRFH(modelSpecs)

  optimizeRE.default(modelSpecs)

  optimizeRE.MSRFH(modelSpecs)
}
\arguments{
  \item{modelSpecs}{modelSpecs object, initialized with
  \code{mS()}}
}
\description{
  Function for estimation of random effects.
}

