% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimizeRho}
\alias{optimizeRho}
\title{optimizeRho}
\usage{
optimizeRho(modelSpecs)
}
\arguments{
\item{modelSpecs}{list with all necessary components for estimation}
}
\description{
find optimal rho coefficients for given variance parameters
}

