% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeInferenceCensusDataSet.R
\name{sae.inference.census}
\alias{sae.inference.census}
\title{SAE prediction on census}
\usage{
sae.inference.census(model, censusdata, location, n_obs_census,
  n_obs_survey, n_locations, welfare.function, n_boot, model_fit_survey,
  inference_survey)
}
\arguments{
\item{x}{Beschreibung von \code{x}}

\item{y}{Beschreibung von \code{y}}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
