% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeInferenceSurveyDataSet.R
\name{sae.inference.survey}
\alias{sae.inference.survey}
\title{Titel der Funktion}
\usage{
sae.inference.survey(model, surveydata, location, id = c())
}
\arguments{
\item{x}{Beschreibung von \code{x}}

\item{y}{Beschreibung von \code{y}}
}
\value{
The function returns a list with
1) a data.frame with
  a) the calculated location effects as average over the residuals in one location
  b) the residuals calculate as residuals = regresson_residuals - location_effect
2) the estimated x'beta from the regression model
}
\description{
Beschreibung der Funktion
}
