% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae overall function.R
\name{sae}
\alias{sae}
\title{SAE main function}
\usage{
sae(model, surveydata, censusdata, location_survey, mResponse,
  n_boot = 50, welfare.function)
}
\arguments{
\item{model}{a model that is specified for the relationship betwenn
the response varibale and the regressors. Model must be a linear model that can be processed by \code{lm()}}

\item{y}{Beschreibung von \code{y}}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
