% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimizeRE}
\alias{optimizeRE}
\alias{optimizeRE.MSRFH}
\alias{optimizeRE.MSSTRFH}
\alias{optimizeRE.default}
\title{optimizeRE - generic funciton}
\usage{
optimizeRE(modelSpecs)

\method{optimizeRE}{MSSTRFH}(modelSpecs)

\method{optimizeRE}{default}(modelSpecs)

\method{optimizeRE}{MSRFH}(modelSpecs)
}
\arguments{
\item{modelSpecs}{modelSpecs object, initialized with \code{mS()}}
}
\description{
Function for estimation of random effects.
}

