% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae overall function.R
\name{sae}
\alias{sae}
\title{SAE main function}
\usage{
sae(model, surveydata, censusdata, location_survey, mResponse,
  n_boot = 50, welfare.function)
}
\arguments{
\item{x}{Beschreibung von \code{x}}

\item{y}{Beschreibung von \code{y}}
}
\value{
Was die Funktion ausspuckt.
}
\description{
Beschreibung der Funktion
}
